Conflictos

+++ app/src/main/java/models/UserId.kt
@@ -1,6 +1,5 @@
-// models/UserId.kt
 package com.example.apprepartidor.models
 
 data class UserId(
-    val user_id: Int
+    val user_id: Long
 )
